chore: Se integraron los servicios del blacklist y del envío de correos mediante el uso de docker-compose
