+++ ServicesLayer/routes/Router.go
@@ -71,7 +71,5 @@ func init(){
 
 func StartServer(){
 	fmt.Println("Server listening on port 8080")
-	router.Run("0.0.0.0:8080")
-	
-	
+	router.Run("0.0.0.0:8080")		
 }
