--- ServicesLayer/services/IMailerService.go
@@ -1,5 +1,0 @@
-package services
-
-type IMailerService interface {
-	SendEmail()
-}
