+++ ServicesLayer/services/mailerService/IMailerService.go
@@ -1,0 +1,5 @@
+package mailerService
+
+type IMailerService interface {
+	SendEmail(emailAddress, token string) (bool)
+}
