+++ helpers/tokenBlackList/TokenBlackList.go
@@ -17,7 +17,7 @@ type tokenBlackList struct { }
 
 func init() {
 	rdb = redis.NewClient(&redis.Options{
-		Addr: "localhost:6379",
+		Addr: "host.docker.internal:6379",
 		Password: "",
 		DB: 0,
 	})
